+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Fixed
+- dice rolling with modified advantage dice for DH characters
+- dice rolling with disadvantage for DH characters
+
 ## [0.3.24] - 2025-12-16
 ### Added
 - talents integration for D&D 2024 characters